requirements.txt
playwright
pandas
openpyxl
plyer
google-generativeai
pydantic
schedule
python-dotenv
fastapi
uvicorn
sqlalchemy
